Stats: Some cleanup of PromoCards wrapper component (#71259)
* Renames the MobilePromoCardWrapper component to PromoCards
* Renames the containing directory to match component name
* Cleans up the display logic and HTML output a bit
