+++ client/my-sites/stats/site.jsx
@@ -39,7 +39,7 @@ import isPrivateSite from 'calypso/state/selectors/is-private-site';
 import { isJetpackSite } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId, getSelectedSiteSlug } from 'calypso/state/ui/selectors';
 import HighlightsSection from './highlights-section';
-import MobilePromoCardWrapper from './mobile-promo-cards';
+import PromoCards from './promo-cards';
 import ChartTabs from './stats-chart-tabs';
 import Countries from './stats-countries';
 import DatePicker from './stats-date-picker';
@@ -367,8 +367,7 @@ class StatsSite extends Component {
 						}
 					</div>
 				</div>
-				{ /** Promo Card is disabled for Odyssey because it doesn't make much sense in the context, which also removes an API call to `plugins`. */ }
-				<MobilePromoCardWrapper isOdysseyStats={ isOdysseyStats } slug={ slug } />
+				<PromoCards isJetpack={ isJetpack } isOdysseyStats={ isOdysseyStats } slug={ slug } />
 				<JetpackColophon />
 			</div>
 		);
